Provide a metadata script
Somehow, in the process of trying to minimize/centralize the
tsconfigs, I managed to get rid of the metadata script and
failed to run a build afterwards.

Still need to reduce duplication, centralize the tsconfigs, but
the current setup works for now, and there is probably a lot
of other cleanup that needs to be done, so this will do for now.

Signed-off-by: Robert Penny <rpenny@rocketsoftware.com>